# docs/swagger.yaml
basePath: /
definitions:
...
  postgres.Problem:
    properties:
      description:
        type: string
      difficulty:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      memory_limit_mb:
        type: integer
      testcases:
        items:
          $ref: '#/definitions/postgres.TestCase'
        type: array
      time_limit_ms:
        type: integer
      title:
        type: string
      user_id:
        type: integer
    type: object
...
  postgres.TestCase:
    properties:
      id:
        type: integer
      input_data:
        type: string
      is_active:
        type: boolean
      output_data:
        type: string
      problem_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  utils.Envelope:
    additionalProperties: {}
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a simple online judge
  title: Adel Online Judge API
  version: "1.0"
paths:
  /admin/problems:
    get:
      consumes:
      - application/json
      description: Get all problems including inactive ones (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                problems:
                  items:
                    $ref: '#/definitions/postgres.Problem'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get all problems for admin
      tags:
      - admin
  /problems:
    get:
      consumes:
      - application/json
      description: Get all active problems
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                problems:
                  items:
                    $ref: '#/definitions/postgres.Problem'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
      summary: Get all problems
      tags:
      - problems
    post:
      consumes:
      - application/json
      description: Create a new problem (admin only)
      parameters:
      - description: Problem object
        in: body
        name: problem
        required: true
        schema:
          $ref: '#/definitions/postgres.Problem'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                problem:
                  $ref: '#/definitions/postgres.Problem'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create problem
      tags:
      - problems
  /problems/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a problem by ID (admin only)
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete problem
      tags:
      - problems
    get:
      consumes:
      - application/json
      description: Get a single problem by its ID
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                problem:
                  $ref: '#/definitions/postgres.Problem'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
      summary: Get problem by ID
      tags:
      - problems
    put:
      consumes:
      - application/json
      description: Update a problem by ID (admin only)
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: integer
      - description: Problem update object
        in: body
        name: problem
        required: true
        schema:
          properties:
            description:
              type: string
            difficulty:
              type: string
            is_active:
              type: boolean
            memory_limit_mb:
              type: integer
            testcases:
              items:
                $ref: '#/definitions/postgres.TestCase'
              type: array
            time_limit_ms:
              type: integer
            title:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                problem:
                  $ref: '#/definitions/postgres.Problem'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update problem
      tags:
      - problems
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"